+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Installation section in README.md
 - Setup section in README.md
 
+### Changed
+
+- Method `changeDatabase()` renamed to `modifyConnection()`
+
 ## [0.1.0] - 2021-02-27
 
 ### Added